second commit

+++ src/main/java/com/sobczynski/bank/model/BankTransfer.java
@@ -2,16 +2,16 @@ package com.sobczynski.bank.model;
 
 public class BankTransfer {
 
-    private double transferId;
+    private Integer transferId;
     private String titleTransfer;
     private double cashTransfer;
-    private double senderId;
-    private double receiverId;
+    private Integer senderId;
+    private Integer receiverId;
 
     public BankTransfer() {
     }
 
-    public BankTransfer(double transferId, String titleTransfer, double cashTransfer, double senderId, double receiverId) {
+    public BankTransfer(Integer transferId, String titleTransfer, double cashTransfer, Integer senderId, Integer receiverId) {
         this.transferId = transferId;
         this.titleTransfer = titleTransfer;
         this.cashTransfer = cashTransfer;
@@ -19,14 +19,14 @@ public class BankTransfer {
         this.receiverId = receiverId;
     }
 
-    public double getTransferId() { return transferId; }
-    public void setTransferId(double transferId) { this.transferId = transferId; }
+    public Integer getTransferId() { return transferId; }
+    public void setTransferId(Integer transferId) { this.transferId = transferId; }
     public String getTitleTransfer() { return titleTransfer; }
     public void setTitleTransfer(String titleTransfer) { this.titleTransfer = titleTransfer; }
     public double getCashTransfer() { return cashTransfer; }
     public void setCashTransfer(double cashTransfer) { this.cashTransfer = cashTransfer; }
-    public double getSenderId() { return senderId; }
-    public void setSenderId(double senderId) { this.senderId = senderId; }
-    public double getReceiverId() { return receiverId; }
-    public void setReceiverId(double receiverId) { this.receiverId = receiverId; }
+    public Integer getSenderId() { return senderId; }
+    public void setSenderId(Integer senderId) { this.senderId = senderId; }
+    public Integer getReceiverId() { return receiverId; }
+    public void setReceiverId(Integer receiverId) { this.receiverId = receiverId; }
 }
